Assignments make and randomly assign.

+++ wac/signals/handlers.py
@@ -18,17 +18,23 @@ def person_delete(sender, instance, **kwargs):
         instance.mugshot.delete(False)
 
 
+
+
+#########################################################################
 # Concerning Weeks and Assignments
+#########################################################################
 
 @receiver(pre_save, sender=Week)
 def obsolete_old_weeks(sender, instance, **kwargs):
-    print("obsolete_old_weeks called")
-    old_weeks = Week.objects.filter(
-        user__exact=instance.user
-    ).update(is_current=False)
+    if not instance.pk:
+        print("obsolete_old_weeks called")
+        old_weeks = Week.objects.filter(
+            user__exact=instance.user
+        ).update(is_current=False)
 
 
 THIS_WEEK = None
+TOTAL_MINUTES_GATHERED = 0
 
 INTERVAL_CHOICES = [
     ('Daily', 'Daily'),
@@ -48,13 +54,13 @@ def create_assignment(chore, date):
         what = chore,
         when = date,
     )
+    global TOTAL_MINUTES_GATHERED
+    TOTAL_MINUTES_GATHERED += chore.duration
     new_assignment.save()
-
 
 def create_repeating_assignments(chore, date, increment):
     print("loop called")
     for d in range(0, 7, increment):
-        print(d)
         create_assignment(chore, date)
         date = date + datetime.timedelta(days=increment)
 
@@ -82,8 +88,6 @@ def weekly(chore):
         number_of_days = randint(0, 1)
     date = THIS_WEEK.start_date + datetime.timedelta(days=number_of_days)
     create_assignment(chore, date)
-
-
 
 def every2weeks(chore):
     print("method every 2 weeks")
@@ -114,27 +118,48 @@ assigning_methods = {
 }
 
 def get_chores_for_this_week():
-    print(THIS_WEEK)
-    print(THIS_WEEK.start_date)
     for chore in Chore.objects.filter(user=THIS_WEEK.user):
         for choice in INTERVAL_CHOICES:
             if chore.interval == choice[0]:
                 method_name = choice[0].replace(' ', '').lower()
                 assigning_methods[method_name](chore)
 
+
 def assign_people_to_chores():
     print('assigning people')
+    people = Person.objects.filter(user__exact=THIS_WEEK.user)
+    to_dos = Assignment.objects.filter(week__exact=THIS_WEEK)
+
+    number_of_people = len(people)
+    for to_do in to_dos:
+        person = randint(0, number_of_people-1)
+        to_do.who = people[person]
+        to_do.save(update_fields=['who'])
+
+
+
 
 @receiver(post_save, sender=Week)
-def week_post_save(sender, instance, **kwargs):
-    print('='*20)
-    print('handlers - week_post_save')
-    print('='*20)
-    global THIS_WEEK
-    THIS_WEEK = instance
+def week_post_save(sender, instance, created, **kwargs):
+    if created:
+        global TOTAL_MINUTES_GATHERED
+        TOTAL_MINUTES_GATHERED = 0
+        print('='*20)
+        print('handlers - week_post_save')
+        print('='*20)
+        global THIS_WEEK
+        THIS_WEEK = instance
 
-    # TEMP - delete all assignments per user.
-    Assignment.objects.exclude(week=THIS_WEEK).delete()
+        # TEMP - delete all assignments per user.
+        Assignment.objects.exclude(week=THIS_WEEK).delete()
 
-    get_chores_for_this_week()
-    assign_people_to_chores()
+        get_chores_for_this_week()
+        assign_people_to_chores()
+
+        THIS_WEEK.is_current = True
+        THIS_WEEK.total_time = TOTAL_MINUTES_GATHERED
+        THIS_WEEK.save()
+        print(Assignment.objects.filter(week=THIS_WEEK))
+
+        # TEMP
+        print(THIS_WEEK.total_time)
